# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FPT Event Management API",
    "version": "1.0.0",
    "description": "API for FPT Event Management System - Go Lambda"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/venues": {
      "get": {
        "tags": [
          "Venues"
        ],
        "summary": "View venue list with nested areas",
        "description": "Returns Venue list with Areas array nested inside",
        "operationId": "getVenues",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of venues with nested areas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Venue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Venues"
        ],
        "summary": "Create new venue (Admin)",
        "operationId": "createVenue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Venues"
        ],
        "summary": "Update venue (Admin)",
        "operationId": "updateVenue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Venues"
        ],
        "summary": "Delete venue (Soft Delete)",
        "operationId": "deleteVenue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Soft deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessage"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/areas/free": {
      "get": {
        "tags": [
          "Venues"
        ],
        "summary": "Staff get free areas with 1h buffer (STAFF/ADMIN)",
        "description": "Check JWT role STAFF/ADMIN, pass startTime and endTime, returns free Venue_Area list with 1 hour buffer.",
        "operationId": "getFreeAreas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start time (ISO 8601, e.g. 2025-12-10T08:00:00)"
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End time (after startTime)"
          }
        ],
        "responses": {
          "200": {
            "description": "Free areas list for Staff to choose",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeAreasResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid startTime/endTime"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not STAFF/ADMIN"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "StatusMessage": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Venue": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "integer"
          },
          "venueName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueArea"
            }
          }
        }
      },
      "VenueCreateRequest": {
        "type": "object",
        "required": [
          "venueName"
        ],
        "properties": {
          "venueName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "VenueUpdateRequest": {
        "type": "object",
        "required": [
          "venueId",
          "venueName",
          "address",
          "status"
        ],
        "properties": {
          "venueId": {
            "type": "integer"
          },
          "venueName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "VenueArea": {
        "type": "object",
        "properties": {
          "areaId": {
            "type": "integer"
          },
          "venueId": {
            "type": "integer"
          },
          "areaName": {
            "type": "string"
          },
          "floor": {
            "type": "integer"
          },
          "capacity": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "FreeAreasResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "bufferHours": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueArea"
            }
          }
        }
      },
...
    }
  }
}